set error to ok if everything is alright, split rerun_c modules
